+++ authentik/rbac/tests/test_api_assigned_by_roles.py
@@ -5,7 +5,7 @@ from rest_framework.test import APITestCase
 from authentik.core.models import Group
 from authentik.core.tests.utils import create_test_admin_user, create_test_user
 from authentik.lib.generators import generate_id
-from authentik.rbac.api.rbac_roles import RoleAssignedObjectPermissionSerializer
+from authentik.rbac.api.rbac_assigned_by_roles import RoleAssignedObjectPermissionSerializer
 from authentik.rbac.models import Role
 from authentik.stages.invitation.models import Invitation
