+++ authentik/rbac/tests/test_api_assigned_by_users.py
@@ -6,7 +6,7 @@ from rest_framework.test import APITestCase
 from authentik.core.models import Group
 from authentik.core.tests.utils import create_test_admin_user, create_test_user
 from authentik.lib.generators import generate_id
-from authentik.rbac.api.rbac_users import UserAssignedObjectPermissionSerializer
+from authentik.rbac.api.rbac_assigned_by_users import UserAssignedObjectPermissionSerializer
 from authentik.rbac.models import Role
 from authentik.stages.invitation.models import Invitation
